Crate rustml

Library for doing machine learning with Rust.

Performance

When it comes to vector and matrix operations rustml makes heavy use of highly optimized numeric libraries like BLAS or ATLAS. By default CBLAS is used because it is installed on many systems by default. However, in many cases performance can be greatly improved when switching to ATLAS. For a detailed description on how to optimize the numeric computations please read the separate documentation on this topic available here.

Example how to do classifications

In the following example a simple k-nearest neighbour algorithm is used to predict the label of a vector with two features based on the examples in the matrix m (the training set) with their known labels stored in the vector labels.

use rustml::*;

let m = mat![  // training set
    1.0, 2.0;  // each row contains one example for which the label is
    1.1, 2.1;  // known
    2.0, 3.0;
    0.9, 1.9;
    2.1, 2.9
];

let labels = vec![1, 2, 2, 1, 2];

// predict the label for feature vector [1.3, 2.0]
let target = 
    knn::classify(
        &m, &labels, &[1.3, 2.0], 
        3, // look at the 3 nearest neighbours to make the decision
        |x, y| Euclid::compute(x, y).unwrap() // use Euclidean distance
    );
assert_eq!(target, 1);

All examples

k-nearest neighbor: Classifies the examples of the test set of the MNIST database of handwritten digits with a simple k-nearest neighbor approach.

matrix multiplication: Multiplies two 6000x6000 matrices.

vector addition: Add vectors.

Reexports

pub use distance::{Distance, Euclid};

pub use matrix::{HasNan, Matrix};

pub use math::{Dimension, Normalization, Mean, Sum, Var};

pub use ops::{MatrixScalarOps, VectorScalarOps, VectorVectorOps};

pub use gaussian::{GaussianEstimator, GaussianFunctions, Gaussian};

Modules

blas
Bindings for BLAS/ATLAS for high performance vector and matrix operations.

consts

csv
Functions to parse and create comma-separated values (CSV).

datasets
Module to easily access popular datasets often used to measure the performance of machine learning algorithms.

distance
Functions to compute the distance between vectors.

gaussian
Module to handle Gaussian distributions.

io
Functions to read and write files (e.g. gzip compressed files).

knn
Functions to compute the k-nearest neighbours.

math
Module with a collection of different mathematical functions.

matrix
Module that contains structs and functions useful for doing matrix operations.

norm
Functions to compute norms of vectors.

ops

vectors
Functions to operate on vectors.

Macros

mat!
Macro to create a matrix.