rustml::blas::cblas_sgemm

pub unsafe extern fn cblas_sgemm(order: Order, transA: Transpose, transB: Transpose, m: c_int, n: c_int, k: c_int, alpha: c_float, A: *const c_float, lda: c_int, B: *const c_float, ldb: c_int, beta: c_float, C: *mut c_float, ldc: c_int)

Computes alpha * op(A) * op(B) + beta * C and stores the result in C.

The parameters alpha and beta are scalars of type f32, A, B, C are a matrices with elements of type f32 and op(X) is either op(X) = X or op(X) = X^T (the transpose or conjugate transpose of the matrix X).

For a high level interface you should use s_gemm in the module ops_inplace.