rustml::blas::cblas_sgemv

pub unsafe extern fn cblas_sgemv(order: Order, transA: Transpose, m: c_int, n: c_int, alpha: c_float, a: *const c_float, lda: c_int, x: *const c_float, incx: c_int, beta: c_float, y: *mut c_float, incy: c_int)

Computes alpha * A * x + beta * y or alpha * A^T * x + beta * y and stores the result in y.

The parameter order specifies the memory layout of the matrix A. Matrices in rustml are stored in RowMajor order by default. If the parameter transA is set to Trans the transpose of A is used, otherwise A. The parameter m specifies the number of rows of A, n the number of columns, lda should be set to the number of columns of A.

For a high level interface you should use s_gemv in the module ops_inplace.