Module rustml::gaussian

Module to handle Gaussian distributions.

Examples

If the parameters of a Gaussian distribution are already known a Gaussian distribution can be simply created as follows.

use rustml::gaussian::*;

let g = Gaussian::new(0.5 /* mean */, 0.1 /* variance */);
assert_eq!(g.mean(), 0.5);
assert_eq!(g.var(), 0.1);
// standard deviation
assert!(g.std() - 0.31623 <= 0.0001);

Structs

Gaussian
Contains all parameters to describe a Gaussian distribution, i.e. the mean and the variance.

Traits

GaussianEstimator
Trait to estimate the mean and the variance of a set of samples.

GaussianFunctions