Struct rustml::gaussian::Gaussian

pub struct Gaussian<T> {
    // some fields omitted
}

Contains all parameters to describe a Gaussian distribution, i.e. the mean and the variance.

Given a set of samples the parameters can be estimated for all types for which the trait Gaussian is implemented.

Methods

impl<T: Float> Gaussian<T>

fn new(mean: T, var: T) -> Gaussian<T>

Creates a gaussian distribution from the given mean mean and the variance var.