Module rustml::io

Functions to read and write files (e.g. gzip compressed files).

Structs

GzipData
Struct to decompress gzip streams.

MatchLines
Read lines from a reader and returns the line if it matches with a regex.

Functions

match_lines
Returns an instance of MatchLines which can be used to read all lines from the given reader. Each line is matched against the provided regex and only those lines which match the regex are returned.

match_lines_stdin
Returns an instance of MatchLines which can be used to read all lines from stdin and match each line with the provided regex. Only those lines are returned which match the regex.