Module rustml::math

Module with a collection of different mathematical functions.

Enums

Dimension
Determines the dimension over which to perform an operation.

Normalization
Determines the type of normalization used for computing the variance or standard deviation.

Traits

Mean
Trait to compute the mean of values.

Sum
Trait to compute the sum of values.

Var
Trait to compute variance of values.