Enum rustml::math::Dimension

pub enum Dimension {
    Row,
    Column,
}

Determines the dimension over which to perform an operation.

Variants

Row

Perform the operation over all elements of a row.

Column

Perform the operatino over all elements of a column.