Struct rustml::matrix::Matrix

pub struct Matrix<T> {
    // some fields omitted
}

A matrix with elements of type T.

An Option is returned which is None if the matrices cannot be multiplied because the number of columns of the left matrix is not equal to the number of rows of the right matrix.

Because the trait Mul is implemented for matrices where T is f32 or f64 such matrices can be multplied with the * operator.

Multiplication of matrices

To multiply two matrices the underlying BLAS implementation is used. By default this is CBLAS. In many cases performance can be greatly improved when switching to ATLAS. For a detailed description on how to optimize the numeric computations please read the separate documentation on this topic available here.

Example

use rustml::*;

let a = mat![
    1.0f32, 5.0, 2.0; 
    2.0, 2.0, 3.0 
];
let b = mat![
    3.0, 7.0, 4.0, 8.0;
    4.0, 2.0, 1.0, 4.0;
    5.0, 2.0, 1.0, 9.0
];
let c = (a * b).unwrap();
assert_eq!(c.row(0).unwrap(), &[33.0, 21.0, 11.0, 46.0]);
assert_eq!(c.row(1).unwrap(), &[29.0, 24.0, 13.0, 51.0]);

Methods

impl<T: Clone> Matrix<T>

fn new() -> Matrix<T>

Creates a new matrix with 0 rows and 0 columns.

use rustml::Matrix;

let m = Matrix::<f32>::new();
assert_eq!(m.rows(), 0);
assert_eq!(m.cols(), 0);

fn fill(value: T, rows: usize, cols: usize) -> Matrix<T>

Creates a matrix with the given number of rows and columns where each element is set to value.

use rustml::Matrix;

let m = Matrix::<f32>::fill(1.2, 2, 2);
assert_eq!(m.row(0).unwrap(), [1.2, 1.2]);
assert_eq!(m.row(1).unwrap(), [1.2, 1.2]);

fn from_vec(vals: Vec<T>, rows: usize, cols: usize) -> Option<Matrix<T>>

Creates a matrix with the given number of rows and columns. The matrix is initialized with the values from the vector vals. The elements are arranged in row-major order in the vector.

use rustml::Matrix;

let m = Matrix::<f32>::from_vec(
    vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0], 3, 2
).unwrap();
assert_eq!(m.row(0).unwrap(), [1.0, 2.0]);
assert_eq!(m.row(1).unwrap(), [3.0, 4.0]);
assert_eq!(m.row(2).unwrap(), [5.0, 6.0]);

fn random<R: Rand + Clone>(rows: usize, cols: usize) -> Matrix<R>

Creates a matrix with random values.

Example

use rustml::Matrix;

let m = Matrix::<f64>::random::<f64>(3, 2);
assert_eq!(m.rows(), 3);
assert_eq!(m.cols(), 2);
// all values are in the interval [0, 1)
assert!(m.values().all(|&x| x < 1.0 && x >= 0.0));
println!("{}", m);

fn buf(&self) -> &Vec<T>

Returns the internal buffer that is used to store the matrix.

fn lead_dim(&self) -> usize

Is equivalent to calling the method cols() on the matrix.

fn rows(&self) -> usize

Returns the number of rows of the matrix.

use rustml::*;

let m = mat![
    1.0, 1.5, 1.8; 
    2.0, 2.5, 2.8
];
assert_eq!(m.rows(), 2);

fn cols(&self) -> usize

Returns the number of columns of the matrix.

use rustml::*;

let m = mat![
    1.0, 1.5, 1.8; 
    2.0, 2.5, 2.8
];
assert_eq!(m.cols(), 3);

fn values(&self) -> Iter<T>

Returns an iterator over all elements of the matrix in row-major order.

use rustml::*;

let m = mat![
    1.0, 1.5; 
    2.0, 2.5
];
let mut i = m.values();
assert_eq!(i.next().unwrap(), &1.0);
assert_eq!(i.next().unwrap(), &1.5);
assert_eq!(i.next().unwrap(), &2.0);

fn row_iter(&self) -> RowIterator<T>

Returns an iterator over the rows of the matrix.

Each call to the next method is done in O(1).

use rustml::*;

let m = mat![
    1.0, 1.5; 
    2.0, 2.5;
    5.0, 5.5
];
let mut i = m.row_iter();
assert_eq!(i.next().unwrap(), [1.0, 1.5]);
assert_eq!(i.next().unwrap(), [2.0, 2.5]);
assert_eq!(i.next().unwrap(), [5.0, 5.5]);

fn row_iter_at(&self, n: usize) -> RowIterator<T>

Returns an iterator over the rows of the matrix. The iterator starts at the the row with index n.

use rustml::*;

let m = mat![
    1.0, 1.5; 
    2.0, 2.5;
    5.0, 5.5
];
let mut i = m.row_iter_at(1);
assert_eq!(i.next().unwrap(), [2.0, 2.5]);
assert_eq!(i.next().unwrap(), [5.0, 5.5]);

fn row_iter_of(&self, rows: &[usize]) -> SelectedRowIterator<T>

Returns an iterator over the rows of the matrix with the specified indexes in rows.

use rustml::*;

let m = mat![
    1.0, 1.5; 
    2.0, 2.5;
    3.0, 3.5;
    4.0, 4.5;
    5.0, 5.5
];
let mut i = m.row_iter_of(&[1, 3, 4]);
assert_eq!(i.next().unwrap(), [2.0, 2.5]);
assert_eq!(i.next().unwrap(), [4.0, 4.5]);
assert_eq!(i.next().unwrap(), [5.0, 5.5]);

fn get(&self, row: usize, col: usize) -> Option<&T>

Returns the element of the matrix at row row (indexing starts at zero) and column col or None if row or column does not exist.

use rustml::*;

let m = mat![
    1.0, 1.5; 
    2.0, 2.5;
    5.0, 5.5
];
assert_eq!(m.get(0, 1).unwrap(), &1.5);
assert_eq!(m.get(2, 0).unwrap(), &5.0);
assert!(m.get(3, 0).is_none());

fn row(&self, n: usize) -> Option<&[T]>

Returns the row at index n (starting at zero) in O(1).

use rustml::*;

let m = mat![
    1.0, 1.5; 
    2.0, 2.5;
    5.0, 5.5
];
assert_eq!(m.row(0).unwrap(), [1.0, 1.5]);
assert_eq!(m.row(1).unwrap(), [2.0, 2.5]);
assert_eq!(m.row(2).unwrap(), [5.0, 5.5]);
assert!(m.row(3).is_none())

fn set(&mut self, row: usize, col: usize, newval: T) -> bool

Replaces the element at row row (indexing starts at zero) and column col with the new value newval. Returns true on success and false on failure, i.e. if the row or column does not exist.

use rustml::*;

let mut m = mat![
    1.0, 1.5; 
    2.0, 2.5;
    5.0, 5.5
];
assert_eq!(m.get(1, 0).unwrap(), &2.0);
m.set(1, 0, 8.0);
assert_eq!(m.get(1, 0).unwrap(), &8.0);
assert_eq!(m.set(3, 0, 8.9), false);

fn map<F, U>(&self, f: F) -> Matrix<U> where F: FnMut(&T) -> U