Trait rustml::matrix::Similar

pub trait Similar<T> {
    fn similar(&self, e: &Self, epsilon: T) -> bool;
}

Trait to check if the values of two matrices with the same dimension differ only within a small range.

Required Methods

fn similar(&self, e: &Self, epsilon: T) -> bool