Struct rustml::nn::NeuralNetwork

pub struct NeuralNetwork {
    // some fields omitted
}

Methods

impl NeuralNetwork

Implementation of a neural network.

fn new() -> NeuralNetwork

Creates a new neural network.

The network does not contain any layer. Layers can be added with the method add_layer.

fn add_layer(&self, n: usize) -> NeuralNetwork

Adds an layer to the neural network.

fn set_params(&self, layer: usize, params: Matrix<f64>) -> NeuralNetwork

Sets the parameters which connect the given layer layer with the next layer.

fn input_size(&self) -> usize

Returns the number of input units.

fn output_size(&self) -> usize

Returns the number of output units.

fn layers(&self) -> usize

Returns the number of layers.

fn error(&self, examples: &Matrix<f64>, targets: &Matrix<f64>) -> f64

Computes the error of the net.

fn predict(&self, input: &[f64]) -> Vec<f64>

Computes the output of the neural network for the given input.

fn optimize(&self, x: &Matrix<f64>, labels: &Matrix<f64>, p: OptParams<f64>) -> NeuralNetwork

fn derivatives(&self, examples: &Matrix<f64>, targets: &Matrix<f64>) -> Vec<Matrix<f64>>

fn update_params(&mut self, deltas: &[Matrix<f64>])

Updates the parameters of the network.

Each matrix in deltas is added to the corresponding matrix of parameters of the network, i.e. the first matrix which contains the parameters from the first layer to the second layer, the second matrix is added to the parameters which contains the parameters from the second layer to the third layer and so on.

fn params(&self) -> Vec<Matrix<f64>>

Returns the parameters of the network.