Trait rustml::ops_inplace::MatrixMatrixOpsInPlace

pub trait MatrixMatrixOpsInPlace<T> {
    fn iadd(&mut self, rhs: &Matrix<T>);
    fn isub(&mut self, rhs: &Matrix<T>);
    fn imule(&mut self, rhs: &Matrix<T>);
}

Trait for matrix-matrix operations.

Required Methods

fn iadd(&mut self, rhs: &Matrix<T>)

Adds the matrix rhs to this matrix inplace.

Implementation details: iterates through the rows of both matrices and uses iadd for the vectors (using BLAS).

fn isub(&mut self, rhs: &Matrix<T>)

Subtracts the matrix rhs from this matrix inplace.

Implementation details: iterates through the rows of both matrices and uses isub for the vectors (using BLAS).

fn imule(&mut self, rhs: &Matrix<T>)

Element wise matrix multiplication.

(not accelerated via BLAS)