Trait rustml::ops_inplace::MatrixScalarOpsInPlace

pub trait MatrixScalarOpsInPlace<T> {
    fn idiv_scalar(&mut self, val: T);
    fn imul_scalar(&mut self, val: T);
    fn iadd_scalar(&mut self, val: T);
    fn isub_scalar(&mut self, val: T);
}

Trait for matrix-scalar operations.

Required Methods

fn idiv_scalar(&mut self, val: T)

Divides each element of the matrix by the scalar val.

Not accelerated via BLAS.

fn imul_scalar(&mut self, val: T)

Multiplies each element of the matrix with the scalar val.

Not accelerated via BLAS.

fn iadd_scalar(&mut self, val: T)

Adds the scalar val to each element of the matrix.

Not accelerated via BLAS.

fn isub_scalar(&mut self, val: T)

Subtracts the scalar val from each element of the matrix.

Not accelerated via BLAS.