Module rustml::regression

Module for linear regression.

Structs

Hypothesis
Hypothesis for linear regression.

Traits

DesignMatrix
Trait to create the design matrix of a matrix of features, i.e. a new column is inserted at the left of the matrix where all elements are equal to one.