Module rustml::vectors

Functions for vectors.

Structs

VecReader
Iterator which reads lines from a reader and converts each line into a vector with elements of the specified type.

Traits

Append

Linspace
Linearly spaced elements.

Select

VectorIO

Functions

copy_memory
Copies elements from src to dst.

from_file
Returns a VecReader to read vectors line by line from a file.

from_reader
Returns a VecReader to read vectors line by line from a reader.

from_value

group
Counts and compresses consecutive elements that are equal.

random
Creates a vector with random elements.

zero
Creates a vector for which all elements are equal to zero.