Module rustml::math

Module with a collection of different mathematical functions.

Reexports

pub use self::sum::{Sum, SumVec};

pub use self::mean::{Mean, MeanVec};

pub use self::var::Var;

Modules

mean

sum

var

Enums

Dimension
Determines the dimension over which to perform an operation.

Normalization
Determines the type of normalization used for computing the variance or standard deviation.