Struct rustml::matrix::Matrix

pub struct Matrix<T> {
    // some fields omitted
}

A matrix with elements of type T.

Methods

impl<T: Clone> Matrix<T>

fn new() -> Matrix<T>

Creates a new matrix with 0 rows and 0 columns.

fn fill(value: T, rows: usize, cols: usize) -> Matrix<T>

Creates a matrix with the given number of rows and columns where each element is set to value.

fn from_vec(vals: Vec<T>, rows: usize, cols: usize) -> Option<Matrix<T>>

Creates a matrix with the given number of rows and columns. The matrix is initialized with the values from the vector vals where the elements are arranged in row-major order in the vector.

fn random<R: Rand + Clone>(rows: usize, cols: usize) -> Matrix<R>

Creates a matrix with random values.

Example

let m = Matrix::<f64>::random::<f64>(3, 2);
assert_eq!(m.rows(), 3);
assert_eq!(m.cols(), 2);
println!("{}", m);

fn lead_dim(&self) -> usize

Is equivalent with calling the method cols.

fn rows(&self) -> usize

Returns the number of rows of the matrix.

fn cols(&self) -> usize

Returns the number of columns of the matrix.

fn values(&self) -> Iter<T>

Returns an iterator over all elements of the matrix in row-major order.

fn row_iter(&self) -> RowIterator<T>

Returns an iterator over the rows of the matrix.

Each call to the next method is done in O(1).

fn row_iter_at(&self, n: usize) -> RowIterator<T>

Returns an iterator over the rows of the matrix where the iterator starts at the the row with index n.

fn row_iter_of(&self, rows: &[usize]) -> SelectedRowIterator<T>

fn get(&self, row: usize, col: usize) -> Option<&T>

Returns the element of the matrix at row row (starting at zero) and column col (starting at zero) or None if row or column does not exist.

fn row(&self, n: usize) -> Option<&[T]>

Returns the row at index n (starting at zero) in O(1).

fn set(&mut self, row: usize, col: usize, newval: T) -> bool

Replaces the element at row row (starting at zero) and column col (starting at zero) with the new value newval. Returns true on success and false on failure, i.e. if the row or column does not exist.